+++ ansible_runner/config/_base.py
@@ -39,6 +39,8 @@ from ansible_runner.exceptions import ConfigurationError
 from ansible_runner.defaults import registry_auth_prefix
 from ansible_runner.loader import ArtifactLoader
 from ansible_runner.utils import (
+    callback_mount,
+    get_callback_dir,
     open_fifo_write,
     args2cmdline,
     sanitize_container_name,
@@ -199,8 +201,6 @@ class BaseConfig(object):
         if self.containerized:
             self.container_name = "ansible_runner_{}".format(sanitize_container_name(self.ident))
             self.env = {}
-            # Special flags to convey info to entrypoint or process in container
-            self.env['LAUNCHED_BY_RUNNER'] = '1'
 
             if self.process_isolation_executable == 'podman':
                 # A kernel bug in RHEL < 8.5 causes podman to use the fuse-overlayfs driver. This results in errors when
@@ -258,17 +258,23 @@ class BaseConfig(object):
         if not self.containerized:
             callback_dir = self.env.get('AWX_LIB_DIRECTORY', os.getenv('AWX_LIB_DIRECTORY'))
             if callback_dir is None:
-                callback_dir = os.path.join(os.path.split(os.path.abspath(__file__))[0], "..", "callbacks")
-            python_path = self.env.get('PYTHONPATH', os.getenv('PYTHONPATH', ''))
-            self.env['PYTHONPATH'] = ':'.join([python_path, callback_dir])
-            if python_path and not python_path.endswith(':'):
-                python_path += ':'
+                callback_dir = get_callback_dir()
             self.env['ANSIBLE_CALLBACK_PLUGINS'] = ':'.join(filter(None, (self.env.get('ANSIBLE_CALLBACK_PLUGINS'), callback_dir)))
 
-        if 'AD_HOC_COMMAND_ID' in self.env:
-            self.env['ANSIBLE_STDOUT_CALLBACK'] = 'minimal'
-        else:
-            self.env['ANSIBLE_STDOUT_CALLBACK'] = 'awx_display'
+        # this is an adhoc command if the module is specified, TODO: combine with logic in RunnerConfig class
+        is_adhoc = bool((getattr(self, 'binary', None) is None) and (getattr(self, 'module', None) is not None))
+
+        if self.env.get('ANSIBLE_STDOUT_CALLBACK'):
+            self.env['ORIGINAL_STDOUT_CALLBACK'] = self.env.get('ANSIBLE_STDOUT_CALLBACK')
+
+        if is_adhoc:
+            # force loading awx_display stdout callback for adhoc commands
+            self.env["ANSIBLE_LOAD_CALLBACK_PLUGINS"] = '1'
+            if 'AD_HOC_COMMAND_ID' not in self.env:
+                self.env['AD_HOC_COMMAND_ID'] = '1'
+
+        self.env['ANSIBLE_STDOUT_CALLBACK'] = 'awx_display'
+
         self.env['ANSIBLE_RETRY_FILES_ENABLED'] = 'False'
         if 'ANSIBLE_HOST_KEY_CHECKING' not in self.env:
             self.env['ANSIBLE_HOST_KEY_CHECKING'] = 'False'
@@ -479,20 +485,18 @@ class BaseConfig(object):
                                             dst_mount_path="/runner/artifacts",
                                             labels=":Z")
 
-            # Mount the entire private_data_dir
-            # custom show paths inside private_data_dir do not make sense
-            self._update_volume_mount_paths(new_args,
-                                            "{}".format(self.private_data_dir),
-                                            dst_mount_path="/runner",
-                                            labels=":Z")
         else:
             subdir_path = os.path.join(self.private_data_dir, 'artifacts')
             if not os.path.exists(subdir_path):
                 os.mkdir(subdir_path, 0o700)
 
-            # Mount the entire private_data_dir
-            # custom show paths inside private_data_dir do not make sense
-            self._update_volume_mount_paths(new_args, "{}".format(self.private_data_dir), dst_mount_path="/runner", labels=":Z")
+        # Mount the entire private_data_dir
+        # custom show paths inside private_data_dir do not make sense
+        self._update_volume_mount_paths(new_args, "{}".format(self.private_data_dir), dst_mount_path="/runner", labels=":Z")
+
+        # Mount the stdout callback plugin from the ansible-runner code base
+        mount_paths = callback_mount(copy_if_needed=True)
+        self._update_volume_mount_paths(new_args, mount_paths[0], dst_mount_path=mount_paths[1], labels=":Z")
 
         if self.container_auth_data:
             # Pull in the necessary registry auth info, if there is a container cred
