--- docs/ansible_runner.callbacks.rst
@@ -1,19 +1,0 @@
-ansible_runner.callbacks package
-================================
-
-Submodules
-----------
-
-ansible_runner.callbacks.awx_display module
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-.. automodule:: ansible_runner.callbacks.awx_display
-    :members:
-    :undoc-members:
-
-ansible_runner.callbacks.minimal module
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-.. automodule:: ansible_runner.callbacks.minimal
-    :members:
-    :undoc-members:
