+++ setup.cfg
@@ -22,3 +22,5 @@ data-files =
 # W503 - Line break occurred before a binary operator
 ignore=W503
 max-line-length=160
+per-file-ignores =
+    ansible_runner/display_callback/callback/awx_display.py:E402
